requirements.txt
asgiref==3.9.1
awscli==1.42.11
boto3==1.40.6
botocore==1.40.11
certifi==2025.7.9
charset-normalizer==3.4.2
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
docutils==0.19
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==25.0
psycopg2-binary==2.9.10
PuLP==3.2.2
pyasn1==0.6.1
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.4
rsa==4.7.2
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
